resolve: Move late resolution visitor into a separate file
